requirements.txt
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html

jax[cuda11_pip]==0.4.19
typing-extensions>=4.7.0
mujoco==2.3.7
cloudpickle==3.0.0
crafter==1.8.3
numpy==1.23.5
scipy==1.11.3
flax==0.7.5
dm-haiku==0.0.12
optax==0.1.9
scikit-learn==1.4.1.post1
seaborn==0.13.2
rich==13.7.1
ruamel.yaml==0.18.6
moviepy==1.0.3
imageio==2.34.0
ipython==8.22.2
ipykernel==6.29.3
hydra-core==1.3.2
wandb==0.16.4
brax==0.9.1
tensorflow-cpu==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.33.0
tensorflow-probability==0.20.1
torch==2.2.1
attrdict==2.0.1
colorlog==6.8.2
evotorch==0.5.1
